+++ src/routes/Profile.js
@@ -4,6 +4,7 @@ import { useNavigate } from "react-router-dom";
 import { collection, orderBy, query, getDocs, getFirestore,doc,updateDoc, toDate } from "firebase/firestore";
 import { updateProfile } from "@firebase/auth";
 import ShowList from "components/showList";
+import ShowBucket from "components/ShowBucket";
 
 
 const Profile=({ userObj })=>{
@@ -38,49 +39,7 @@ const Profile=({ userObj })=>{
             window.location.reload();
         }
     };
-    useEffect(async()=>{
-        const bucketsRef=collection(getFirestore(),"users/"+userObj.uid+"/buckets");
-        //const q = query(bucketsRef, orderBy('dateAt','desc'));
-        const temp=await getDocs(bucketsRef);
-        const tempBuckets=[];
-        temp.forEach((doc)=>{
-            var created=new Date(doc.data().dateAt);
-            created=created.getFullYear()+'-'+created.getMonth()+'-'+created.getDate();
-            tempBuckets.push(
-                {
-                    'id':doc.id,
-                    'text':doc.data().text,
-                    'dateAt':created,
-                    'expiredAt':doc.data().expiredAt,
-                    'tags':doc.data().tags,
-                    'userId':doc.data().userId,
-                }
-            );
-        });
-        
-        setBuckets(tempBuckets);
-    },[]);
-    //console.log(userObj.displayName);
-
-    const [isChecked, setIsChecked] = useState(false);
-    const[checkedbuckets, setCheckedbuckets] = useState(new Set());
-
-    const checkHandler = ({ target }) => {
-        setIsChecked(!isChecked);
-        checkedbucketHandler(target.parentNode, target.value, target.checked);
-    };
-
-    const checkedbucketHandler = ( id, isChecked) => {
-        if(isChecked) {
-            checkedbuckets.add(id);
-            setCheckedbuckets(checkedbuckets);
-        }
-        else if (!isChecked && checkedbuckets.has(id)) {
-            checkedbuckets.delete(id);
-            setCheckedbuckets(checkedbuckets);
-        }
-        return checkedbuckets;
-    };
+   
 
     return (
 
@@ -104,24 +63,7 @@ const Profile=({ userObj })=>{
                 Log Out
             </span>
             
-            <div>
-                {buckets.map((bucket) => (
-                    <div key={bucket.id}>
-                        <div>{bucket.text}</div>
-                        <div>{bucket.tags}</div>
-                        <div>created {bucket.dateAt}</div>
-                        <div>expired {bucket.expiredAt}</div>
-                        <div>completed
-                            <label className="innerBox">
-                                <input
-                                    type="checkbox"
-                                    value={buckets.text}
-                                    onChange={(e) => checkHandler(e)} />
-                            </label>
-                        </div>
-                    </div>
-                ))}
-            </div>
+            <ShowBucket userObj={userObj}/>
         </div>
     );
